+++ core/domain/exp_domain.py
@@ -22,20 +22,16 @@ should therefore be independent of the specific storage models used.
 """
 
 import copy
-import logging
 import re
 import string
 
 from constants import constants
-from core.domain import html_cleaner
 from core.domain import html_validation_service
 from core.domain import interaction_registry
 from core.domain import param_domain
-import core.domain.state_domain as state_domain
+from core.domain import state_domain
 from core.platform import models
 import feconf
-import jinja_utils
-import schema_utils
 import utils
 
 (exp_models,) = models.Registry.import_models([models.NAMES.exploration])
@@ -95,105 +91,6 @@ STATISTICAL_CLASSIFICATION = 'statistical_classifier'
 # Represents answers which led to the 'default outcome' of an interaction,
 # rather than belonging to a specific answer group.
 DEFAULT_OUTCOME_CLASSIFICATION = 'default_outcome'
-
-
-def get_full_customization_args(customization_args, ca_specs):
-    """Populates the given customization_args dict with default values
-    if any of the expected customization_args are missing.
-
-    Args:
-        customization_args: dict. The customization dict. The keys are names of
-            customization_args and the values are dicts with a
-            single key, 'value', whose corresponding value is the value of
-            the customization arg.
-        ca_specs: list(dict). List of spec dictionaries. Is used to check if
-            some keys are missing in customization_args. Dicts have the
-            following structure:
-                - name: str. The customization variable name.
-                - description: str. The customization variable description.
-                - default_value: *. The default value of the customization
-                    variable.
-
-    Returns:
-        dict. The customization_args dict where missing keys are populated with
-        the default values.
-    """
-    for ca_spec in ca_specs:
-        if ca_spec.name not in customization_args:
-            customization_args[ca_spec.name] = {
-                'value': ca_spec.default_value
-            }
-    return customization_args
-
-
-def validate_customization_args_and_values(
-        item_name, item_type, customization_args,
-        ca_specs_to_validate_against):
-    """Validates the given `customization_args` dict against the specs set out
-    in 'ca_specs_to_validate_against'. 'item_name' and 'item_type' are used to
-    populate any error messages that arise during validation.
-    Note that this may modify the given customization_args dict, if it has
-    extra or missing keys. It also normalizes any HTML in the
-    customization_args dict.
-
-    Args:
-        item_name: str. This is always 'interaction'.
-        item_type: str. The item_type is the ID of the interaction.
-        customization_args: dict. The customization dict. The keys are names of
-            customization_args and the values are dicts with a
-            single key, 'value', whose corresponding value is the value of
-            the customization arg.
-        ca_specs_to_validate_against: list(dict). List of spec dictionaries. Is
-            used to check if some keys are missing in customization_args. Dicts
-            have the following structure:
-                - name: str. The customization variable name.
-                - description: str. The customization variable description.
-                - default_value: *. The default value of the customization
-                    variable.
-
-    Raises:
-        ValidationError: The given 'customization_args' is not valid.
-    """
-    ca_spec_names = [
-        ca_spec.name for ca_spec in ca_specs_to_validate_against]
-
-    if not isinstance(customization_args, dict):
-        raise utils.ValidationError(
-            'Expected customization args to be a dict, received %s'
-            % customization_args)
-
-    # Validate and clean up the customization args.
-
-    # Populate missing keys with the default values.
-    customization_args = get_full_customization_args(
-        customization_args, ca_specs_to_validate_against)
-
-    # Remove extra keys.
-    extra_args = []
-    for arg_name in customization_args.keys():
-        if not isinstance(arg_name, basestring):
-            raise utils.ValidationError(
-                'Invalid customization arg name: %s' % arg_name)
-        if arg_name not in ca_spec_names:
-            extra_args.append(arg_name)
-            logging.warning(
-                '%s %s does not support customization arg %s.'
-                % (item_name.capitalize(), item_type, arg_name))
-    for extra_arg in extra_args:
-        del customization_args[extra_arg]
-
-    # Check that each value has the correct type.
-    for ca_spec in ca_specs_to_validate_against:
-        try:
-            customization_args[ca_spec.name]['value'] = (
-                schema_utils.normalize_against_schema(
-                    customization_args[ca_spec.name]['value'],
-                    ca_spec.schema))
-        except Exception:
-            # TODO(sll): Raise an actual exception here if parameters are not
-            # involved (If they are, can we get sample values for the state
-            # context parameters?).
-            pass
 
 
 class ExplorationChange(object):
@@ -381,92 +278,6 @@ class ExplorationCommitLogEntry(object):
         }
 
 
-class AudioTranslation(object):
-    """Value object representing an audio translation."""
-
-    def to_dict(self):
-        """Returns a dict representing this AudioTranslation domain object.
-
-        Returns:
-            dict. A dict, mapping all fields of AudioTranslation instance.
-        """
-        return {
-            'filename': self.filename,
-            'file_size_bytes': self.file_size_bytes,
-            'needs_update': self.needs_update,
-        }
-
-    @classmethod
-    def from_dict(cls, audio_translation_dict):
-        """Return a AudioTranslation domain object from a dict.
-
-        Args:
-            audio_translation_dict: dict. The dict representation of
-                AudioTranslation object.
-
-        Returns:
-            AudioTranslation. The corresponding AudioTranslation domain object.
-        """
-        return cls(
-            audio_translation_dict['filename'],
-            audio_translation_dict['file_size_bytes'],
-            audio_translation_dict['needs_update'])
-
-    def __init__(self, filename, file_size_bytes, needs_update):
-        """Initializes a AudioTranslation domain object.
-
-        Args:
-            filename: str. The corresponding audio file path.
-            file_size_bytes: int. The file size, in bytes. Used to display
-                potential bandwidth usage to the learner before they download
-                the file.
-            needs_update: bool. Whether audio is marked for needing review.
-        """
-        # str. The corresponding audio file path, e.g.
-        # "content-en-2-h7sjp8s.mp3".
-        self.filename = filename
-        # int. The file size, in bytes. Used to display potential bandwidth
-        # usage to the learner before they download the file.
-        self.file_size_bytes = file_size_bytes
-        # bool. Whether audio is marked for needing review.
-        self.needs_update = needs_update
-
-    def validate(self):
-        """Validates properties of the AudioTranslation.
-
-        Raises:
-            ValidationError: One or more attributes of the AudioTranslation are
-            invalid.
-        """
-        if not isinstance(self.filename, basestring):
-            raise utils.ValidationError(
-                'Expected audio filename to be a string, received %s' %
-                self.filename)
-        dot_index = self.filename.rfind('.')
-        if dot_index == -1 or dot_index == 0:
-            raise utils.ValidationError(
-                'Invalid audio filename: %s' % self.filename)
-        extension = self.filename[dot_index + 1:]
-        if extension not in feconf.ACCEPTED_AUDIO_EXTENSIONS:
-            raise utils.ValidationError(
-                'Invalid audio filename: it should have one of '
-                'the following extensions: %s. Received: %s'
-                % (feconf.ACCEPTED_AUDIO_EXTENSIONS.keys(), self.filename))
-
-        if not isinstance(self.file_size_bytes, int):
-            raise utils.ValidationError(
-                'Expected file size to be an int, received %s' %
-                self.file_size_bytes)
-        if self.file_size_bytes <= 0:
-            raise utils.ValidationError(
-                'Invalid file size: %s' % self.file_size_bytes)
-
-        if not isinstance(self.needs_update, bool):
-            raise utils.ValidationError(
-                'Expected needs_update to be a bool, received %s' %
-                self.needs_update)
-
-
 class ExpVersionReference(object):
     """Value object representing an exploration ID and a version number."""
 
@@ -506,806 +317,6 @@ class ExpVersionReference(object):
         if not isinstance(self.version, int):
             raise utils.ValidationError(
                 'Expected version to be an int, received %s' % self.version)
-
-
-class SubtitledHtml(object):
-    """Value object representing subtitled HTML."""
-
-    def __init__(self, content_id, html):
-        """Initializes a SubtitledHtml domain object.
-
-        Args:
-            content_id: str. A unique id referring to the audio translations for
-              this content.
-            html: str. A piece of user submitted HTML. This is cleaned in such
-                a way as to contain a restricted set of HTML tags.
-        """
-        self.content_id = content_id
-        self.html = html_cleaner.clean(html)
-        self.validate()
-
-    def to_dict(self):
-        """Returns a dict representing this SubtitledHtml domain object.
-
-        Returns:
-            dict. A dict, mapping all fields of SubtitledHtml instance.
-        """
-        return {
-            'content_id': self.content_id,
-            'html': self.html
-        }
-
-    @classmethod
-    def from_dict(cls, subtitled_html_dict):
-        """Return a SubtitledHtml domain object from a dict.
-
-        Args:
-            subtitled_html_dict: dict. The dict representation of SubtitledHtml
-                object.
-
-        Returns:
-            SubtitledHtml. The corresponding SubtitledHtml domain object.
-        """
-        return cls(
-            subtitled_html_dict['content_id'], subtitled_html_dict['html'])
-
-    def validate(self):
-        """Validates properties of the SubtitledHtml.
-
-        Raises:
-            ValidationError: One or more attributes of the SubtitledHtml are
-            invalid.
-        """
-        if not isinstance(self.content_id, basestring):
-            raise utils.ValidationError(
-                'Expected content id to be a string, received %s' %
-                self.content_id)
-
-        # TODO(sll): Add HTML sanitization checking.
-        # TODO(sll): Validate customization args for rich-text components.
-        if not isinstance(self.html, basestring):
-            raise utils.ValidationError(
-                'Invalid content HTML: %s' % self.html)
-
-    def to_html(self, params):
-        """Exports this SubtitledHTML object to an HTML string. The HTML is
-        parameterized using the parameters in `params`.
-
-        Args:
-            params: dict. The keys are the parameter names and the values are
-                the values of parameters.
-
-        Raises:
-            Exception: 'params' is not a dict.
-
-        Returns:
-            str. The HTML string that results after stripping
-                out unrecognized tags and attributes.
-        """
-        if not isinstance(params, dict):
-            raise Exception(
-                'Expected context params for parsing subtitled HTML to be a '
-                'dict, received %s' % params)
-
-        return html_cleaner.clean(jinja_utils.parse_string(self.html, params))
-
-    @classmethod
-    def create_default_subtitled_html(cls, content_id):
-        """Create a default SubtitledHtml domain object."""
-        return cls(content_id, '')
-
-
-class RuleSpec(object):
-    """Value object representing a rule specification."""
-
-    def to_dict(self):
-        """Returns a dict representing this RuleSpec domain object.
-
-        Returns:
-            dict. A dict, mapping all fields of RuleSpec instance.
-        """
-        return {
-            'rule_type': self.rule_type,
-            'inputs': self.inputs,
-        }
-
-    @classmethod
-    def from_dict(cls, rulespec_dict):
-        """Return a RuleSpec domain object from a dict.
-
-        Args:
-            rulespec_dict: dict. The dict representation of RuleSpec object.
-
-        Returns:
-            RuleSpec. The corresponding RuleSpec domain object.
-        """
-        return cls(
-            rulespec_dict['rule_type'],
-            rulespec_dict['inputs']
-        )
-
-    def __init__(self, rule_type, inputs):
-        """Initializes a RuleSpec domain object.
-
-        Args:
-            rule_type: str. The rule type, e.g. "CodeContains" or "Equals". A
-                full list of rule types can be found in
-                extensions/interactions/rule_templates.json.
-            inputs: dict. The values of the parameters needed in order to fully
-                specify the rule. The keys for this dict can be deduced from
-                the relevant description field in
-                extensions/interactions/rule_templates.json -- they are
-                enclosed in {{...}} braces.
-        """
-        self.rule_type = rule_type
-        self.inputs = inputs
-
-    def validate(self, rule_params_list, exp_param_specs_dict):
-        """Validates a RuleSpec value object. It ensures the inputs dict does
-        not refer to any non-existent parameters and that it contains values
-        for all the parameters the rule expects.
-
-        Args:
-            rule_params_list: A list of parameters used by the rule represented
-                by this RuleSpec instance, to be used to validate the inputs of
-                this RuleSpec. Each element of the list represents a single
-                parameter and is a tuple with two elements:
-                    0: The name (string) of the parameter.
-                    1: The typed object instance for that
-                        parameter (e.g. Real).
-            exp_param_specs_dict: A dict of specified parameters used in this
-                exploration. Keys are parameter names and values are ParamSpec
-                value objects with an object type property (obj_type). RuleSpec
-                inputs may have a parameter value which refers to one of these
-                exploration parameters.
-
-        Raises:
-            ValidationError: One or more attributes of the RuleSpec are
-            invalid.
-        """
-        if not isinstance(self.inputs, dict):
-            raise utils.ValidationError(
-                'Expected inputs to be a dict, received %s' % self.inputs)
-        input_key_set = set(self.inputs.keys())
-        param_names_set = set([rp[0] for rp in rule_params_list])
-        leftover_input_keys = input_key_set - param_names_set
-        leftover_param_names = param_names_set - input_key_set
-
-        # Check if there are input keys which are not rule parameters.
-        if leftover_input_keys:
-            logging.warning(
-                'RuleSpec \'%s\' has inputs which are not recognized '
-                'parameter names: %s' % (self.rule_type, leftover_input_keys))
-
-        # Check if there are missing parameters.
-        if leftover_param_names:
-            raise utils.ValidationError(
-                'RuleSpec \'%s\' is missing inputs: %s'
-                % (self.rule_type, leftover_param_names))
-
-        rule_params_dict = {rp[0]: rp[1] for rp in rule_params_list}
-        for (param_name, param_value) in self.inputs.iteritems():
-            param_obj = rule_params_dict[param_name]
-            # Validate the parameter type given the value.
-            if isinstance(param_value, basestring) and '{{' in param_value:
-                # Value refers to a parameter spec. Cross-validate the type of
-                # the parameter spec with the rule parameter.
-                start_brace_index = param_value.index('{{') + 2
-                end_brace_index = param_value.index('}}')
-                param_spec_name = param_value[
-                    start_brace_index:end_brace_index]
-                if param_spec_name not in exp_param_specs_dict:
-                    raise utils.ValidationError(
-                        'RuleSpec \'%s\' has an input with name \'%s\' which '
-                        'refers to an unknown parameter within the '
-                        'exploration: %s' % (
-                            self.rule_type, param_name, param_spec_name))
-                # TODO(bhenning): The obj_type of the param_spec
-                # (exp_param_specs_dict[param_spec_name]) should be validated
-                # to be the same as param_obj.__name__ to ensure the rule spec
-                # can accept the type of the parameter.
-            else:
-                # Otherwise, a simple parameter value needs to be normalizable
-                # by the parameter object in order to be valid.
-                param_obj.normalize(param_value)
-
-
-class Outcome(object):
-    """Value object representing an outcome of an interaction. An outcome
-    consists of a destination state, feedback to show the user, and any
-    parameter changes.
-    """
-
-    def to_dict(self):
-        """Returns a dict representing this Outcome domain object.
-
-        Returns:
-            dict. A dict, mapping all fields of Outcome instance.
-        """
-        return {
-            'dest': self.dest,
-            'feedback': self.feedback.to_dict(),
-            'labelled_as_correct': self.labelled_as_correct,
-            'param_changes': [
-                param_change.to_dict() for param_change in self.param_changes],
-            'refresher_exploration_id': self.refresher_exploration_id,
-            'missing_prerequisite_skill_id': self.missing_prerequisite_skill_id
-        }
-
-    @classmethod
-    def from_dict(cls, outcome_dict):
-        """Return a Outcome domain object from a dict.
-
-        Args:
-            outcome_dict: dict. The dict representation of Outcome object.
-
-        Returns:
-            Outcome. The corresponding Outcome domain object.
-        """
-        return cls(
-            outcome_dict['dest'],
-            SubtitledHtml.from_dict(outcome_dict['feedback']),
-            outcome_dict['labelled_as_correct'],
-            [param_domain.ParamChange(
-                param_change['name'], param_change['generator_id'],
-                param_change['customization_args'])
-             for param_change in outcome_dict['param_changes']],
-            outcome_dict['refresher_exploration_id'],
-            outcome_dict['missing_prerequisite_skill_id']
-        )
-
-    def __init__(
-            self, dest, feedback, labelled_as_correct, param_changes,
-            refresher_exploration_id, missing_prerequisite_skill_id):
-        """Initializes a Outcome domain object.
-
-        Args:
-            dest: str. The name of the destination state.
-            feedback: SubtitledHtml. Feedback to give to the user if this rule
-                is triggered.
-            labelled_as_correct: bool. Whether this outcome has been labelled
-                by the creator as corresponding to a "correct" answer.
-            param_changes: list(ParamChange). List of exploration-level
-                parameter changes to make if this rule is triggered.
-            refresher_exploration_id: str or None. An optional exploration ID
-                to redirect the learner to if they seem to lack understanding
-                of a prerequisite concept. This should only exist if the
-                destination state for this outcome is a self-loop.
-            missing_prerequisite_skill_id: str or None. The id of the skill that
-                this answer group tests. If this is not None, the exploration
-                player would redirect to this skill when a learner receives this
-                outcome.
-        """
-        # Id of the destination state.
-        # TODO(sll): Check that this state actually exists.
-        self.dest = dest
-        # Feedback to give the reader if this rule is triggered.
-        self.feedback = feedback
-        # Whether this outcome has been labelled by the creator as
-        # corresponding to a "correct" answer.
-        self.labelled_as_correct = labelled_as_correct
-        # Exploration-level parameter changes to make if this rule is
-        # triggered.
-        self.param_changes = param_changes or []
-        # An optional exploration ID to redirect the learner to if they lack
-        # understanding of a prerequisite concept. This should only exist if
-        # the destination state for this outcome is a self-loop.
-        self.refresher_exploration_id = refresher_exploration_id
-        # An optional skill id whose concept card would be shown to the learner
-        # when the learner receives this outcome.
-        self.missing_prerequisite_skill_id = missing_prerequisite_skill_id
-
-    def validate(self):
-        """Validates various properties of the Outcome.
-
-        Raises:
-            ValidationError: One or more attributes of the Outcome are invalid.
-        """
-        self.feedback.validate()
-
-        if not isinstance(self.labelled_as_correct, bool):
-            raise utils.ValidationError(
-                'The "labelled_as_correct" field should be a boolean, received '
-                '%s' % self.labelled_as_correct)
-
-        if self.missing_prerequisite_skill_id is not None:
-            if not isinstance(self.missing_prerequisite_skill_id, basestring):
-                raise utils.ValidationError(
-                    'Expected outcome missing_prerequisite_skill_id to be a '
-                    'string, received %s' % self.missing_prerequisite_skill_id)
-
-        if not isinstance(self.param_changes, list):
-            raise utils.ValidationError(
-                'Expected outcome param_changes to be a list, received %s'
-                % self.param_changes)
-        for param_change in self.param_changes:
-            param_change.validate()
-
-        if self.refresher_exploration_id is not None:
-            if not isinstance(self.refresher_exploration_id, basestring):
-                raise utils.ValidationError(
-                    'Expected outcome refresher_exploration_id to be a string, '
-                    'received %s' % self.refresher_exploration_id)
-
-
-class AnswerGroup(object):
-    """Value object for an answer group. Answer groups represent a set of rules
-    dictating whether a shared feedback should be shared with the user. These
-    rules are ORed together. Answer groups may also support a classifier
-    that involve soft matching of answers to a set of training data and/or
-    example answers dictated by the creator.
-    """
-
-    def to_dict(self):
-        """Returns a dict representing this AnswerGroup domain object.
-
-        Returns:
-            dict. A dict, mapping all fields of AnswerGroup instance.
-        """
-        return {
-            'rule_specs': [rule_spec.to_dict()
-                           for rule_spec in self.rule_specs],
-            'outcome': self.outcome.to_dict(),
-            'training_data': self.training_data,
-            'tagged_misconception_id': self.tagged_misconception_id
-        }
-
-    @classmethod
-    def from_dict(cls, answer_group_dict):
-        """Return a AnswerGroup domain object from a dict.
-
-        Args:
-            answer_group_dict: dict. The dict representation of AnswerGroup
-                object.
-
-        Returns:
-            AnswerGroup. The corresponding AnswerGroup domain object.
-        """
-        return cls(
-            Outcome.from_dict(answer_group_dict['outcome']),
-            [RuleSpec.from_dict(rs) for rs in answer_group_dict['rule_specs']],
-            answer_group_dict['training_data'],
-            answer_group_dict['tagged_misconception_id']
-        )
-
-    def __init__(
-            self, outcome, rule_specs, training_data, tagged_misconception_id):
-        """Initializes a AnswerGroup domain object.
-
-        Args:
-            outcome: Outcome. The outcome corresponding to the answer group.
-            rule_specs: list(RuleSpec). List of rule specifications.
-            training_data: list(*). List of answers belonging to training
-                data of this answer group.
-            tagged_misconception_id: int or None. The id of the tagged
-                misconception for the answer group, when a state is part of a
-                Question object that tests a particular skill.
-        """
-        self.rule_specs = [RuleSpec(
-            rule_spec.rule_type, rule_spec.inputs
-        ) for rule_spec in rule_specs]
-
-        self.outcome = outcome
-        self.training_data = training_data
-        self.tagged_misconception_id = tagged_misconception_id
-
-    def validate(self, interaction, exp_param_specs_dict):
-        """Verifies that all rule classes are valid, and that the AnswerGroup
-        only has one classifier rule.
-
-        Args:
-            exp_param_specs_dict: dict. A dict of all parameters used in the
-                exploration. Keys are parameter names and values are ParamSpec
-                value objects with an object type property (obj_type).
-            interaction: InteractionInstance. The interaction object.
-
-        Raises:
-            ValidationError: One or more attributes of the AnswerGroup are
-                invalid.
-            ValidationError: The AnswerGroup contains more than one classifier
-                rule.
-        """
-        if not isinstance(self.rule_specs, list):
-            raise utils.ValidationError(
-                'Expected answer group rules to be a list, received %s'
-                % self.rule_specs)
-
-        if self.tagged_misconception_id is not None:
-            if not isinstance(self.tagged_misconception_id, int):
-                raise utils.ValidationError(
-                    'Expected tagged misconception id to be an int, '
-                    'received %s' % self.tagged_misconception_id)
-
-        if len(self.rule_specs) == 0 and len(self.training_data) == 0:
-            raise utils.ValidationError(
-                'There must be at least one rule or training data for each'
-                ' answer group.')
-
-        for rule_spec in self.rule_specs:
-            if rule_spec.rule_type not in interaction.rules_dict:
-                raise utils.ValidationError(
-                    'Unrecognized rule type: %s' % rule_spec.rule_type)
-
-            rule_spec.validate(
-                interaction.get_rule_param_list(rule_spec.rule_type),
-                exp_param_specs_dict)
-
-        self.outcome.validate()
-
-
-class Hint(object):
-    """Value object representing a hint."""
-
-    def __init__(self, hint_content):
-        """Constructs a Hint domain object.
-
-        Args:
-            hint_content: SubtitledHtml. The hint text and ID referring to the
-              audio translations for this content.
-        """
-        self.hint_content = hint_content
-
-    def to_dict(self):
-        """Returns a dict representing this Hint domain object.
-
-        Returns:
-            dict. A dict mapping the field of Hint instance.
-        """
-        return {
-            'hint_content': self.hint_content.to_dict(),
-        }
-
-    @classmethod
-    def from_dict(cls, hint_dict):
-        """Return a Hint domain object from a dict.
-
-        Args:
-            hint_dict: dict. The dict representation of Hint object.
-
-        Returns:
-            Hint. The corresponding Hint domain object.
-        """
-        return cls(SubtitledHtml.from_dict(hint_dict['hint_content']))
-
-    def validate(self):
-        """Validates all properties of Hint."""
-        self.hint_content.validate()
-
-
-class Solution(object):
-    """Value object representing a solution.
-
-    A solution consists of answer_is_exclusive, correct_answer and an
-    explanation.When answer_is_exclusive is True, this indicates that it is
-    the only correct answer; when it is False, this indicates that it is one
-    possible answer. correct_answer records an answer that enables the learner
-    to progress to the next card and explanation is an HTML string containing
-    an explanation for the solution.
-    """
-
-    def __init__(
-            self, interaction_id, answer_is_exclusive,
-            correct_answer, explanation):
-        """Constructs a Solution domain object.
-
-        Args:
-            interaction_id: str. The interaction id.
-            answer_is_exclusive: bool. True if is the only correct answer;
-                False if is one of possible answer.
-            correct_answer: str. The correct answer; this answer enables the
-                learner to progress to the next card.
-            explanation: SubtitledHtml. Contains text and text id to link audio
-                translations for the solution's explanation.
-        """
-        self.answer_is_exclusive = answer_is_exclusive
-        self.correct_answer = (
-            interaction_registry.Registry.get_interaction_by_id(
-                interaction_id).normalize_answer(correct_answer))
-        self.explanation = explanation
-
-    def to_dict(self):
-        """Returns a dict representing this Solution domain object.
-
-        Returns:
-            dict. A dict mapping all fields of Solution instance.
-        """
-        return {
-            'answer_is_exclusive': self.answer_is_exclusive,
-            'correct_answer': self.correct_answer,
-            'explanation': self.explanation.to_dict(),
-        }
-
-    @classmethod
-    def from_dict(cls, interaction_id, solution_dict):
-        """Return a Solution domain object from a dict.
-
-        Args:
-            interaction_id: str. The interaction id.
-            solution_dict: dict. The dict representation of Solution object.
-
-        Returns:
-            Solution. The corresponding Solution domain object.
-        """
-        return cls(
-            interaction_id,
-            solution_dict['answer_is_exclusive'],
-            interaction_registry.Registry.get_interaction_by_id(
-                interaction_id).normalize_answer(
-                    solution_dict['correct_answer']),
-            SubtitledHtml.from_dict(solution_dict['explanation']))
-
-    def validate(self, interaction_id):
-        """Validates all properties of Solution.
-
-        Args:
-            interaction_id: str. The interaction id.
-
-        Raises:
-            ValidationError: One or more attributes of the Solution are not
-            valid.
-        """
-        if not isinstance(self.answer_is_exclusive, bool):
-            raise utils.ValidationError(
-                'Expected answer_is_exclusive to be bool, received %s' %
-                self.answer_is_exclusive)
-        interaction_registry.Registry.get_interaction_by_id(
-            interaction_id).normalize_answer(self.correct_answer)
-        self.explanation.validate()
-
-
-class InteractionInstance(object):
-    """Value object for an instance of an interaction."""
-
-    # The default interaction used for a new state.
-    _DEFAULT_INTERACTION_ID = None
-
-    def to_dict(self):
-        """Returns a dict representing this InteractionInstance domain object.
-
-        Returns:
-            dict. A dict mapping all fields of InteractionInstance instance.
-        """
-        return {
-            'id': self.id,
-            'customization_args': (
-                {} if self.id is None else
-                get_full_customization_args(
-                    self.customization_args,
-                    interaction_registry.Registry.get_interaction_by_id(
-                        self.id).customization_arg_specs)),
-            'answer_groups': [group.to_dict() for group in self.answer_groups],
-            'default_outcome': (
-                self.default_outcome.to_dict()
-                if self.default_outcome is not None
-                else None),
-            'confirmed_unclassified_answers': (
-                self.confirmed_unclassified_answers),
-            'hints': [hint.to_dict() for hint in self.hints],
-            'solution': self.solution.to_dict() if self.solution else None,
-        }
-
-    @classmethod
-    def from_dict(cls, interaction_dict):
-        """Return a InteractionInstance domain object from a dict.
-
-        Args:
-            interaction_dict: dict. The dict representation of
-                InteractionInstance object.
-
-        Returns:
-            InteractionInstance. The corresponding InteractionInstance domain
-            object.
-        """
-        default_outcome_dict = (
-            Outcome.from_dict(interaction_dict['default_outcome'])
-            if interaction_dict['default_outcome'] is not None else None)
-        solution_dict = (
-            Solution.from_dict(
-                interaction_dict['id'], interaction_dict['solution'])
-            if interaction_dict['solution'] else None)
-
-        return cls(
-            interaction_dict['id'],
-            interaction_dict['customization_args'],
-            [AnswerGroup.from_dict(h)
-             for h in interaction_dict['answer_groups']],
-            default_outcome_dict,
-            interaction_dict['confirmed_unclassified_answers'],
-            [Hint.from_dict(h) for h in interaction_dict['hints']],
-            solution_dict)
-
-    def __init__(
-            self, interaction_id, customization_args, answer_groups,
-            default_outcome, confirmed_unclassified_answers, hints, solution):
-        """Initializes a InteractionInstance domain object.
-
-        Args:
-            interaction_id: str. The interaction id.
-            customization_args: dict. The customization dict. The keys are
-                names of customization_args and the values are dicts with a
-                single key, 'value', whose corresponding value is the value of
-                the customization arg.
-            answer_groups: list(AnswerGroup). List of answer groups of the
-                interaction instance.
-            default_outcome: Outcome. The default outcome of the interaction
-                instance.
-            confirmed_unclassified_answers: list(AnswerGroup). List of answers
-                which have been confirmed to be associated with the default
-                outcome.
-            hints: list(Hint). List of hints for this interaction.
-            solution: Solution. A possible solution for the question asked in
-                this interaction.
-        """
-        self.id = interaction_id
-        # Customization args for the interaction's view. Parts of these
-        # args may be Jinja templates that refer to state parameters.
-        # This is a dict: the keys are names of customization_args and the
-        # values are dicts with a single key, 'value', whose corresponding
-        # value is the value of the customization arg.
-        self.customization_args = customization_args
-        self.answer_groups = answer_groups
-        self.default_outcome = default_outcome
-        self.confirmed_unclassified_answers = confirmed_unclassified_answers
-        self.hints = hints
-        self.solution = solution
-
-    @property
-    def is_terminal(self):
-        """Determines if this interaction type is terminal. If no ID is set for
-        this interaction, it is assumed to not be terminal.
-
-        Returns:
-            bool. Whether the interaction is terminal.
-        """
-        return self.id and interaction_registry.Registry.get_interaction_by_id(
-            self.id).is_terminal
-
-    def get_all_outcomes(self):
-        """Returns a list of all outcomes of this interaction, taking into
-        consideration every answer group and the default outcome.
-
-        Returns:
-            list(Outcome). List of all outcomes of this interaction.
-        """
-        outcomes = []
-        for answer_group in self.answer_groups:
-            outcomes.append(answer_group.outcome)
-        if self.default_outcome is not None:
-            outcomes.append(self.default_outcome)
-        return outcomes
-
-    def validate(self, exp_param_specs_dict):
-        """Validates various properties of the InteractionInstance.
-
-        Args:
-            exp_param_specs_dict: dict. A dict of specified parameters used in
-                the exploration. Keys are parameter names and values are
-                ParamSpec value objects with an object type property(obj_type).
-                Is used to validate AnswerGroup objects.
-
-        Raises:
-            ValidationError: One or more attributes of the InteractionInstance
-            are invalid.
-        """
-        if not isinstance(self.id, basestring):
-            raise utils.ValidationError(
-                'Expected interaction id to be a string, received %s' %
-                self.id)
-        try:
-            interaction = interaction_registry.Registry.get_interaction_by_id(
-                self.id)
-        except KeyError:
-            raise utils.ValidationError('Invalid interaction id: %s' % self.id)
-
-        validate_customization_args_and_values(
-            'interaction', self.id, self.customization_args,
-            interaction.customization_arg_specs)
-
-        if not isinstance(self.answer_groups, list):
-            raise utils.ValidationError(
-                'Expected answer groups to be a list, received %s.'
-                % self.answer_groups)
-        if not self.is_terminal and self.default_outcome is None:
-            raise utils.ValidationError(
-                'Non-terminal interactions must have a default outcome.')
-        if self.is_terminal and self.default_outcome is not None:
-            raise utils.ValidationError(
-                'Terminal interactions must not have a default outcome.')
-        if self.is_terminal and self.answer_groups:
-            raise utils.ValidationError(
-                'Terminal interactions must not have any answer groups.')
-
-        for answer_group in self.answer_groups:
-            answer_group.validate(interaction, exp_param_specs_dict)
-        if self.default_outcome is not None:
-            self.default_outcome.validate()
-
-        if not isinstance(self.hints, list):
-            raise utils.ValidationError(
-                'Expected hints to be a list, received %s'
-                % self.hints)
-        for hint in self.hints:
-            hint.validate()
-
-        if self.solution:
-            self.solution.validate(self.id)
-
-        if self.solution and not self.hints:
-            raise utils.ValidationError(
-                'Hint(s) must be specified if solution is specified')
-
-    @classmethod
-    def create_default_interaction(cls, default_dest_state_name):
-        """Create a default InteractionInstance domain object:
-            - customization_args: empty dictionary;
-            - answer_groups: empty list;
-            - default_outcome: dest is set to 'default_dest_state_name' and
-                feedback and param_changes are initialized as empty lists;
-            - confirmed_unclassified_answers: empty list;
-
-        Args:
-            default_dest_state_name: str. The default destination state.
-
-        Returns:
-            InteractionInstance. The corresponding InteractionInstance domain
-            object with default values.
-        """
-        default_outcome = Outcome(
-            default_dest_state_name,
-            SubtitledHtml.create_default_subtitled_html(
-                feconf.DEFAULT_OUTCOME_CONTENT_ID), False, {}, None, None)
-
-        return cls(
-            cls._DEFAULT_INTERACTION_ID, {}, [], default_outcome, [], [], {})
-
-
-    def get_all_html_content_strings(self):
-        """Get all html content strings in the interaction.
-
-        Returns:
-            list(str): The list of all html content strings in the interaction.
-        """
-        html_list = []
-
-        for answer_group in self.answer_groups:
-            outcome_html = answer_group.outcome.feedback.html
-            html_list = html_list + [outcome_html]
-
-        # Note that ItemSelectionInput replicates the customization arg HTML
-        # in its answer groups.
-        if self.id == 'ItemSelectionInput':
-            for answer_group in self.answer_groups:
-                for rule_spec in answer_group.rule_specs:
-                    rule_spec_html = rule_spec.inputs['x']
-                    html_list = html_list + rule_spec_html
-
-        if self.id == 'DragAndDropSortInput':
-            for answer_group in self.answer_groups:
-                for rule_spec in answer_group.rule_specs:
-                    rule_spec_html_list = rule_spec.inputs['x']
-                    for rule_spec_html in rule_spec_html_list:
-                        html_list = html_list + rule_spec_html
-
-        if self.default_outcome:
-            default_outcome_html = self.default_outcome.feedback.html
-            html_list = html_list + [default_outcome_html]
-
-        for hint in self.hints:
-            hint_html = hint.hint_content.html
-            html_list = html_list + [hint_html]
-
-        if self.solution:
-            solution_html = self.solution.explanation.html
-            html_list = html_list + [solution_html]
-
-        if self.id in (
-                'ItemSelectionInput', 'MultipleChoiceInput',
-                'DragAndDropSortInput'):
-            customization_args_html_list = (
-                self.customization_args['choices']['value'])
-            html_list = html_list + customization_args_html_list
-
-        return html_list
 
 
 class ExplorationVersionsDiff(object):
@@ -1531,7 +542,7 @@ class Exploration(object):
         for (state_name, sdict) in exploration_dict['states'].iteritems():
             state = exploration.states[state_name]
 
-            state.content = SubtitledHtml(
+            state.content = state_domain.SubtitledHtml(
                 sdict['content']['content_id'], sdict['content']['html'])
 
             state.param_changes = [param_domain.ParamChange(
@@ -1546,27 +557,27 @@ class Exploration(object):
 
             idict = sdict['interaction']
             interaction_answer_groups = [
-                AnswerGroup.from_dict(group)
+                state_domain.AnswerGroup.from_dict(group)
                 for group in idict['answer_groups']]
 
             default_outcome = (
-                Outcome.from_dict(idict['default_outcome'])
+                state_domain.Outcome.from_dict(idict['default_outcome'])
                 if idict['default_outcome'] is not None else None)
 
             solution = (
-                Solution.from_dict(idict['id'], idict['solution'])
+                state_domain.Solution.from_dict(idict['id'], idict['solution'])
                 if idict['solution'] else None)
 
-            state.interaction = InteractionInstance(
+            state.interaction = state_domain.InteractionInstance(
                 idict['id'], idict['customization_args'],
                 interaction_answer_groups, default_outcome,
                 idict['confirmed_unclassified_answers'],
-                [Hint.from_dict(h) for h in idict['hints']],
+                [state_domain.Hint.from_dict(h) for h in idict['hints']],
                 solution)
 
             state.content_ids_to_audio_translations = {
                 content_id: {
-                    language_code: AudioTranslation.from_dict(
+                    language_code: state_domain.AudioTranslation.from_dict(
                         audio_translation_dict)
                     for language_code, audio_translation_dict in
                     audio_translations.iteritems()
